Fix wrapping issue with codespans in MD, extract hint overlay to store+layout
